feat(backend): 재고 수량 타입 변경(Long) 및 목록 조회 시 재고 정보 추가

+++ backend/src/main/java/com/imjjh/Dibs/api/product/dto/response/ProductDetailResponseDto.java
@@ -4,25 +4,25 @@ import lombok.Builder;
 
 @Builder
 public record ProductDetailResponseDto(
-        Long id,
+                Long id,
 
-        String sellerName,
+                String sellerName,
 
-        String name,
+                String name,
 
-        String description,
+                String description,
 
-        Long price,
+                Long price,
 
-        Integer stockQuantity,
+                Long stockQuantity,
 
-        String status,
+                String status,
 
-        String imageUrl,
+                String imageUrl,
 
-        String category,
+                String category,
 
-        Long specialPrice,
+                Long specialPrice,
 
-        Integer discountRate) {
+                Integer discountRate) {
 }
